requirements.txt
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
